requirements.txt
# Core ML & Data
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# MLOps & Tracking
mlflow==2.8.1
optuna==3.4.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=1.10.0,<2.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
scipy==1.11.4
liac-arff==2.5.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Model Interpretation
shap==0.43.0
lime==0.2.0.1

# Monitoring & Quality
evidently==0.4.11
great-expectations==0.18.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
joblib==1.3.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.6.0
